pyproject.toml
[tool.poetry]
name = "whitehamster26_gendiff"
version = "0.3.1"
description = "Show the difference between two config files"
authors = ["whitehamster26"]
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.6"
PyYAML = "^5.3.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"
pytest = "^5.4.1"
pytest-cov = "^2.8.1"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff_run:main"
gendiff_tests = "gendiff.scripts.tests_run:main"

[[tool.poetry.source]]
name = "whitehamster26_gendiff"
url = "https://test.pypi.org/legacy/"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

